requirements.txt
nextflow
seqkit
bbmap
CAT
DeepVirFinder
R
python3
pandas
click
biopython
virsorter
cd-hit
VIBRANT
vContact2
Demovir
